+++ rmb-sdk-go/common/connection.go
@@ -1,4 +1,4 @@
-package direct
+package common
 
 import (
 	"context"
