+++ rmb-sdk-go/common/twindb.go
@@ -1,4 +1,4 @@
-package direct
+package common
 
 import (
 	"fmt"
